+++ packages/chakra-admin/src/core/list/PaginationType.ts
@@ -1,5 +1,0 @@
-export type PaginationType = {
-  limit?: number
-  offset?: number
-  total?: number
-}