+++ tooling/ca-test-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # ca-test-utils
 
+## 0.0.2-dev.12
+
+### Patch Changes
+
+- WIP: Connection Cursor Support
+
 ## 0.0.2-dev.11
 
 ### Patch Changes